// packages/destination-actions/src/destinations/ortto-audiences/syncAudience/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * Hidden field used to verify that the payload is generated by an Audience. Payloads not containing computation_class = 'audience' or 'journey_step' will be dropped before the perform() fuction call.
   */
  segment_computation_action: string
  /**
   * The name of a boolean trait/property in Segment that indicates a user's membership in a specific audience. This name matches Segment's friendly name for the audience exactly.
   */
  computation_key: string
  /**
   * Unique Audience Identifier returned by the createAudience() function call.
   */
  external_audience_id: string
  /**
   * Message ID
   */
  message_id?: string
  /**
   * When enabled, events will be sent to Ortto in batches for improved efficiency.
   */
  enable_batching?: boolean
  /**
   * Maximum number of events to include in each batch.
   */
  batch_size?: number
  /**
   * The unique user identifier
   */
  user_id?: string
  /**
   * Anonymous user identifier
   */
  anonymous_id?: string
  /**
   * The contact's IP address
   */
  ip?: string | null
  /**
   * The contact's location. Will take priority over the IP address.
   */
  location?: {
    country?: string | null
    state?: string | null
    city?: string | null
    post_code?: string | null
  }
  /**
   * A computed object for track and identify events. This field should not need to be edited.
   */
  traits: {
    [k: string]: unknown
  }
}